fix: split cache no-store guards
Separates error response cache protection from sensitive route cache protection and normalizes sensitive paths before matching.
